review: keep the new methods with the types they hang off
AlterWithRenamedCheckConstraints and CheckConstraintsReferenced move into
statement.go with the rest of the AbstractStatement methods, and
newTableAlter and checkConstraintRenames into change.go with the rest of
tableChange's, rather than each pair sitting in a check_constraints.go of
its own. No behaviour change.
